pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comp3065-pet-cw"
version = "0.2.0"
description = "COMP3065 coursework code: Oxford-IIIT Pet classification with a custom CNN and transfer learning."
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "numpy>=1.24",
  "matplotlib>=3.7",
  "scikit-learn>=1.3",
  "seaborn>=0.13",
  "pillow>=10.0",
  "tqdm>=4.66",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
